requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
decorator==4.4.2
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-cloud-core==2.3.2
google-cloud-speech==2.16.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
moviepy==1.0.3
mutagen==1.46.0
numpy==1.23.5
oauthlib==3.2.2
pexels-api-py==0.0.5
Pillow==9.3.0
proglog==0.1.10
proto-plus==1.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.25.1
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
soupsieve==2.3.2.post1
srt==3.5.2
tqdm==4.64.1
uritemplate==4.1.1
urllib3==1.26.12
firebase-admin==5.2.0
demucs==4.0.0
google-cloud-texttospeech